09.04.2023  06.30

+++ searchengine/src/main/java/searchengine/model/Lemma.java
@@ -23,7 +23,7 @@ public class Lemma {
     private String lemma;
 
     @Column(nullable = false)
-    private Long frequency;
+    private int frequency;
 }
